Remove redundant split
