+++ snn_hfo_ieeg/stages/snn/basic_network_creation.py
@@ -1,11 +1,14 @@
 from functools import reduce
-from brian2 import SpikeGeneratorGroup
-from brian2.units import us, second
 import numpy as np
+from brian2 import SpikeGeneratorGroup
+from brian2.units import us, second, amp
+from teili.models.builder.neuron_equation_builder import NeuronEquationBuilder
+from teili.models.builder.synapse_equation_builder import SynapseEquationBuilder
+from teili.core.groups import Neurons, Connections
 from snn_hfo_ieeg.functions.signal_to_spike import concatenate_spikes
 from snn_hfo_ieeg.stages.snn.tau_generation import generate_concatenated_taus
 from snn_hfo_ieeg.stages.snn.weight_generation import generate_weights
-from snn_hfo_ieeg.stages.snn.creation import create_synapses
+from snn_hfo_ieeg.functions.dynapse_biases import get_current
 
 
 def _append_spikes(spikes, spike_train):
@@ -44,3 +47,23 @@ def create_hidden_to_output_synapses(name, hidden_layer, output_layer, model_pat
     taus = np.repeat(10, hidden_neuron_count.hidden)
     return create_synapses(
         f'{name}_hidden_to_{name}_output', model_paths, hidden_layer, output_layer, weights, taus)
+
+
+def create_non_input_layer(model_paths, neuron_count, name, num_inputs=1):
+    equation_builder = NeuronEquationBuilder.import_eq(
+        model_paths.neuron, num_inputs)
+    return Neurons(
+        N=neuron_count,
+        equation_builder=equation_builder,
+        name=f'{name}_layer',
+        dt=100*us)
+
+
+def create_synapses(name, model_paths, from_layer, to_layer, weights, taus):
+    equation_builder = SynapseEquationBuilder.import_eq(model_paths.synapse)
+    synapses = Connections(
+        from_layer, to_layer, equation_builder=equation_builder, name=f'{name}_synapses', verbose=False, dt=100*us)
+    synapses.connect()
+    synapses.weight = weights
+    synapses.I_tau = get_current(taus*1e-3) * amp
+    return synapses
